requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0a1
-e git+https://github.com/mszczesniak89/django-bootstrap-modal-forms.git@6383f67f923e64b7c1581bd5cbe6cd7ee8a547f3#egg=django_bootstrap_modal_forms
django-braces==1.14.0
django-crispy-forms==1.13.0
django-filter==21.1
Faker==9.2.0
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
whitenoise==5.3.0
